rename ZeHardware package and zeclasses
